Replace a bunch of size_t with unsigned
I've apparently gotten into the wrong mindset about size_t vs
unsigned.  It's easy to let size_t proliferate throughout your
code base once you start matching types with all of the standard
library arguments and thinking of this as the "natural" type for
anything that counts or sizes things.  It seems to have a certain
cleanliness advantage in that it won't wrap in the real world,
too.

However, it's commonly wasteful (e.g. x86_64) when the sizes are
known to be reasonably-limited, and better to stick with unsigned
in the general case.

We've kept size_t in places where it clearly is a good idea: when
it's directly storing a size_t return value from a function before
checking and converting, for our own library functions related to
raw allocation and file-mapping, and for file data offsets in
readers of potentially-large files (like the zonefile scanner).

+++ src/cookie.c
@@ -251,7 +251,7 @@ void cookie_config(const char* key_file)
 
     if (key_file) {
         if (safe_read_keyfile(key_file, master_key))
-            log_fatal("Cannot read %zu bytes from '%s': %s", (size_t)KDF_KEYBYTES, key_file, logf_errno());
+            log_fatal("Cannot read %u bytes from '%s': %s", (unsigned)KDF_KEYBYTES, key_file, logf_errno());
     } else {
         char* autokey_path = gdnsd_resolve_path_run(base_autokey, NULL);
         if (safe_read_keyfile(autokey_path, master_key)) {
@@ -284,7 +284,7 @@ void cookie_runtime_init(struct ev_loop* loop)
     ev_periodic_start(loop, hourly_p);
 }
 
-bool cookie_process(uint8_t* cookie_data_out, const uint8_t* cookie_data_in, const gdnsd_anysin_t* client, const size_t cookie_data_in_len)
+bool cookie_process(uint8_t* cookie_data_out, const uint8_t* cookie_data_in, const gdnsd_anysin_t* client, const unsigned cookie_data_in_len)
 {
     gdnsd_assume(master_key); // cookie_config() was run
